Added support for POSIX-based consoles on Windows.
This allows the package to be used on consoles such as those provided by MSYS
and Cygwin. Because these systems don't use the standard cmdhost.exe
interpreter, they won't be detected as terminals by the
golang.org/x/crypto/ssh/terminal package, and methods to switch them to raw mode
won't work.

This change uses the stty command to identify terminals, save the terminal
state, switch the terminal to raw mode, and switch it back. It is reasonably
fast. Unfortunately this is the only way to do this, because Cygwin and MSYS
generally rely on a shim library to translate ioctl/termios calls to work on
their terminals, and Go doesn't have any good way to get at these libraries
(they can't be dynamically loaded because their path is unknown and they can't
be statically linked because that would make the binary far less distributable).