all: remove CDDL-licensed code and modernize package
This package is technically no longer supported, and we should rewrite
it in its entirety in Mutagen, but for now we'll keep it going with a
little facelift and modernization. This includes removing the Travis CI
configuration, adding Go module support, removing the Solaris-specific
code (which was a roadblock for us due to its CDDL licensing), and
replacing the code for all platforms with the newer Go term package.

Signed-off-by: Jacob Howard <jacob@mutagen.io>